add jar file

+++ Lab2_MaxT/src/com/company/View.java
@@ -1,7 +1,7 @@
 package com.company;
 
 public class View {
-    public  void printText(String  str , double sum)
+    public  void printText(String  str , int sum)
     {
     System.out.println("Text : "+str);
     System.out.println("The sum of the numbers in the given text = " + sum);
